Initial import
Case ZC-386:  We need to create RPMs for the ionCube Loader.  Since we've
got multiple versions, and spent so much time getting all the macros in
place to have SCL-building stuff to do the right thing, we can just build
all the packages we need out of this one repo.