Makefile to enable single-repo building of all targets
Case ZC-386:  We need to be able to build all relevant SCL targets out of
a single repository.  The specfile is OK, but unfortunately making OBS
understand SCL is going to be a project well outside the scope of cPanel.
Instead, we can get make to do the work for us, and prepend a couple of
macros onto the specfile for each required version before feeding it back
up to OBS, so OBS will pull in the correct build dependencies.