Add 32-bit targets
Case ZC-807:  It turns out that we do need to do 32-bit targets after all,
so we're adding the 32-bit archive, and some conditional business within
the spec to get the right files in the resulting packages.

Also turned off the debuginfo package because there is almost certainly no
debug symbols in these object files.